refactor: 更新 App 路由位置
