+++ src/BootstrapBlazor.Docs/App.razor
@@ -1,9 +1,9 @@
 ﻿@inject IStringLocalizer<App> AppLocalizer
-@inherits WebSiteModuleComponentBase
-@attribute [JSModuleAutoLoader("App.razor.js", AutoInvokeDispose = false)]
+@inherits BootstrapModuleComponentBase
+@attribute [JSModuleAutoLoader("./_content/BootstrapBlazor.Docs/App.razor.js", AutoInvokeDispose = false)]
 
 <BootstrapBlazorRoot>
-    <Router AppAssembly="@typeof(App).Assembly">
+    <Router AppAssembly="@typeof(App).Assembly" AdditionalAssemblies="AdditionalAssemblies">
         <Found Context="routeData">
             <PageTitle>@AppLocalizer["Title"]</PageTitle>
             <RouteView RouteData="@routeData" DefaultLayout="@typeof(MainLayout)" />
