+++ src/BootstrapBlazor.Docs/App.razor.js
