+++ src/BootstrapBlazor.Docs/BootstrapBlazor.Docs.csproj
@@ -4,14 +4,28 @@
 
   <PropertyGroup>
     <Version>7.0.0</Version>
+    <RootNamespace>BootstrapBlazor.Shared</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
-    <SupportedPlatform Include="browser" />
+    <PackageReference Include="BootstrapBlazor.Shared" Version="7.0.0" />
+    <PackageReference Include="Longbow.GiteeAuth" Version="7.0.0" />
+    <PackageReference Include="Longbow.GitHubAuth" Version="7.0.1" />
+    <PackageReference Include="Longbow.OAuth" Version="7.0.0" />
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\BootstrapBlazor\BootstrapBlazor.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Using Include="BootstrapBlazor.Components" />
+    <Using Include="BootstrapBlazor.Shared.Common" />
+    <Using Include="BootstrapBlazor.Shared.Components" />
+    <Using Include="BootstrapBlazor.Shared.Samples" />
+    <Using Include="Microsoft.AspNetCore.Components" />
+    <Using Include="Microsoft.Extensions.Localization" />
+    <Using Include="System.ComponentModel.DataAnnotations" />
   </ItemGroup>
 
 </Project>
