+++ src/BootstrapBlazor.Docs/Demos/AzureOpenAI/AzureOpenAIChat.razor.css
