+++ src/BootstrapBlazor.Docs/Extensions/DispatchEntryExtensions.cs
