+++ src/BootstrapBlazor.Docs/OAuth/AzureOpenAIUser.cs
