+++ src/BootstrapBlazor.Docs/_Imports.razor
@@ -1,0 +1,18 @@
+﻿@using BootstrapBlazor.Components
+@using BootstrapBlazor.Shared
+@using BootstrapBlazor.Shared.Components
+@using BootstrapBlazor.Shared.Samples
+@using BootstrapBlazor.Shared.Shared
+@using Microsoft.AspNetCore.Authorization
+@using Microsoft.AspNetCore.Components.Forms
+@using Microsoft.AspNetCore.Components.Routing
+@using Microsoft.AspNetCore.Components.Web
+@using Microsoft.Extensions.Configuration
+@using Microsoft.Extensions.DependencyInjection
+@using Microsoft.Extensions.Localization
+@using Microsoft.Extensions.Options
+@using Microsoft.JSInterop
+@using System.Diagnostics.CodeAnalysis
+@using System.Net.Http
+@using System.Threading.Tasks
+@using System.Globalization
