+++ src/BootstrapBlazor.Shared/BootstrapBlazor.Shared.csproj
@@ -2,6 +2,10 @@
 
   <Import Project="..\..\Framework.props" />
 
+  <PropertyGroup>
+    <Version>7.0.0</Version>
+  </PropertyGroup>
+
   <ItemGroup>
     <SupportedPlatform Include="browser" />
   </ItemGroup>
@@ -22,6 +26,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <PackageReference Include="BootstrapBlazor" Version="7.8.3" />
     <PackageReference Include="BootstrapBlazor.AzureOpenAI" Version="7.0.0" />
     <PackageReference Include="BootstrapBlazor.BaiduSpeech" Version="7.1.0" />
     <PackageReference Include="BootstrapBlazor.BaiduOcr" Version="7.1.1" />
@@ -44,13 +49,6 @@
     <PackageReference Include="BootstrapBlazor.Topology" Version="7.4.4" />
     <PackageReference Include="BootstrapBlazor.VideoPlayer" Version="7.0.4" />
     <PackageReference Include="BootstrapBlazor.MouseFollower" Version="7.0.0" />
-    <PackageReference Include="Longbow.GiteeAuth" Version="7.0.0" />
-    <PackageReference Include="Longbow.GitHubAuth" Version="7.0.1" />
-    <PackageReference Include="Longbow.OAuth" Version="7.0.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\BootstrapBlazor.Docs\BootstrapBlazor.Docs.csproj" />
   </ItemGroup>
 
   <ItemGroup>
