+++ src/BootstrapBlazor.Shared/Components/Header.razor.js
@@ -1,4 +1,4 @@
-﻿import EventHandler from "../../../_content/BootstrapBlazor/modules/event-handler.js?v=$version"
+﻿import EventHandler from "../../../_content/BootstrapBlazor/modules/event-handler.js?v=7.8.3"
 
 export function init() {
     const scrollTop = () => (document.documentElement && document.documentElement.scrollTop) || document.body.scrollTop
