+++ src/BootstrapBlazor.Shared/Components/Pre.razor.js
@@ -1,6 +1,6 @@
-﻿import { copy, getDescribedElement, addLink, addScript, getHeight } from "../../BootstrapBlazor/modules/utility.js?v=$version"
-import Data from "../../BootstrapBlazor/modules/data.js?v=$version"
-import EventHandler from "../../BootstrapBlazor/modules/event-handler.js?v=$version"
+﻿import { copy, getDescribedElement, addLink, addScript, getHeight } from "../../BootstrapBlazor/modules/utility.js?v=7.8.3"
+import Data from "../../BootstrapBlazor/modules/data.js?v=7.8.3"
+import EventHandler from "../../BootstrapBlazor/modules/event-handler.js?v=7.8.3"
 
 export async function init(id, title) {
     const el = document.getElementById(id);
