+++ src/BootstrapBlazor.Shared/Components/ThemeChooser.razor.js
@@ -1,6 +1,6 @@
-﻿import { insertAfter } from "../../../_content/BootstrapBlazor/modules/utility.js?v=$version"
-import Data from "../../../_content/BootstrapBlazor/modules/data.js?v=$version"
-import EventHandler from "../../../_content/BootstrapBlazor/modules/event-handler.js?v=$version"
+﻿import { insertAfter } from "../../../_content/BootstrapBlazor/modules/utility.js?v=7.8.3"
+import Data from "../../../_content/BootstrapBlazor/modules/data.js?v=7.8.3"
+import EventHandler from "../../../_content/BootstrapBlazor/modules/event-handler.js?v=7.8.3"
 
 export function init(id) {
     const el = document.getElementById(id)
