+++ src/BootstrapBlazor.Shared/Locales/en.json
@@ -6,7 +6,6 @@
     "PinteresoSummary": "Pintereso"
   },
   "BootstrapBlazor.Shared.App": {
-    "Title": "Bootstrap Blazor enterprise-level UI component library",
     "Reload": "Reload",
     "ErrorMessage": "An unhandled exception has occurred. See browser dev tools for details."
   },
@@ -18,6 +17,7 @@
     "CopiedText": "Copied"
   },
   "BootstrapBlazor.Shared.Pages.Index": {
+    "Title": "Bootstrap Blazor enterprise-level UI component library",
     "BodyClassString": "welcome-body-en",
     "HeaderTitle": "Bootstrap style Blazor UI component library",
     "HeaderDesc": "Based on the <code>Bootstrap</code> style library, it is carefully built, and {0} a variety of commonly used components have been added to bring you an extraordinary feeling for rapid development projects",
