+++ src/BootstrapBlazor.Shared/Locales/zh.json
@@ -18,6 +18,7 @@
     "CopiedText": "复制代码成功"
   },
   "BootstrapBlazor.Shared.Pages.Index": {
+    "Title": "Bootstrap Blazor - 组件库",
     "BodyClassString": "welcome-body",
     "HeaderTitle": "Bootstrap 风格的 Blazor UI 组件库",
     "HeaderDesc": "基于 <code>Bootstrap</code> 样式库精心打造，并且额外增加了 {0} 多种常用的组件，为您快速开发项目带来非一般的感觉",
