+++ src/BootstrapBlazor.Shared/Shared/PracNavMenu.razor.cs
@@ -17,7 +17,7 @@ public partial class PracNavMenu
 {
     [Inject]
     [NotNull]
-    private IStringLocalizer<App>? AppLocalizer { get; set; }
+    private IStringLocalizer<Index>? IndexLocalizer { get; set; }
 
     [Inject]
     [NotNull]
@@ -129,7 +129,7 @@ public partial class PracNavMenu
     {
         if (!item.Items.Any() && !string.IsNullOrEmpty(item.Text))
         {
-            await TitleService.SetTitle($"{item.Text} - {AppLocalizer["Title"]}");
+            await TitleService.SetTitle($"{item.Text} - {IndexLocalizer["Title"]}");
         }
     }
